Create gate.h
From Lab2, the gates now have an array with pointers to their input wires, and a pointer to its output wire. 
The inherited gates have similar functionality. 
The difference from Lab2 is that a gate now has a name, which will be needed later when reading from a txt file.